Experimental use of total body cost as spawning threshold
The value of a creep depends on body parts configuration, so it might be
better to use body cost instead of simply counting number of creeps
and keep it constant.  It's a bit difficult to determine how much cost
should be appropriate as the threshold, but let's try with half the whole
available energy.