Prevent harvesters from picking up from sink links even if harvesting source
Since all action options are compared in terms of walk cost, it might seem
rewarding to harvesters to pick up from sink links whatever other options
are.  Usually it is, but we don't want harvesters to go back to base
cluster everytime someone puts energy into a link.

This crude assumption is that the source and links must be separated more
than 10 squares because you won't build links so close to each other.