Harvesters and builders prefer dropped resources over containers
Becaues dropped resources degrade over time, they need to get it as fast
as possible.  But we also don't like them to go to a dropped resource 10
times further than a container next to them.  So we adjust the cost value
so they do not prefer too much.

I'm not sure if it actually works; it's not easy to test effect of cost
preferences because there are not much occasions that all the other
conditions are the same.