Record average of CPU usage over non-recorded ticks
Because CPU usage greatly varies over ticks, it would introduce much error
if we just record instantaneous values.  Luckily we can measure and average
(moving average, to be exact) of CPU usage and record it instead of
instantaneous values, in order to reduce the errors.

After all, we have CPU bucket, so we don't have much concern on every tick.