Prevent harvesters from moving to another room
It seemed reasonable to move to another room as helping hand if there
is nothing to do in current room, but actually it caused cluttering and
unnecessarily frequent spawning in the original room.  When we exploit a
room, we might be better to have a rule to move workers into the new
room to speed up establishment, but it will be another story.